[Update] アンフォローメソッド

+++ app/models/user.rb
@@ -12,17 +12,12 @@ class User < ApplicationRecord
     withdrawn:            false    #退会済みユーザー
    }
   
-  #フォローしている人を取得するための記述
+  #フォローしているユーザーとのアソシエーション
   has_many :relationships, dependent: :destroy
-  
-  #followingsクラスを仮で指定。relationshipテーブルのfollow_idから、フォローしている人たちを取得
   has_many :followings, through: :relationships, source: :follow
   
-  #フォロワーを取得する記述
-  #reverse_of_relationshipsというクラスを仮で指定。relationsipテーブルのfollow_idを基に、アクセス
+  #フォロワーとのアソシエーション  
   has_many :reverse_of_relationships, class_name: "Relationship", foreign_key: "follow_id", dependent: :destroy
-
-  #先ほど指定したreberse_of_relationshipsを中間テーブルとし、フォロワーのuser_idを取得する
   has_many :followers, through: :reverse_of_relationships, source: :user
 
   #すでにフォローしていればtrueを返す
@@ -40,9 +35,7 @@ class User < ApplicationRecord
   #フォローがあるか確認し、あればアンフォロー
   def unfollow(other_user)
     relationship = self.relationships.find_by(follow_id: other_user.id)
-    if !relationship.empty?
-      relationship.destroy
-    end
+    relationship.destroy if relationship
   end
 
 end
